Create index.html
As you see in the preview post we share you a fully responsive registration form in html5 and css3. Now today’s we share fully responsive price table with html5 and css3. The purpose of this content to show your user to the price table on your website. In this template, we used css3 properties to make the mobile-friendly layout and this is 100% fully responsive.  Here we used a background image with fixed CSS properties box-sizing properties.